--- src/components/Form/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Form';
