+++ src/components/common/index.ts
@@ -5,3 +5,4 @@ export * from './Modal';
 export * from './Button';
 export * from './ErrorBoundary';
 export * from './Input';
+export * from './Form';
